feat: delete protection + publicHomePage done

+++ backend/node_api/src/controllers/commentControllers.ts
@@ -1,6 +1,7 @@
 import { Request, Response } from 'express';
 import * as CommentService from '../services/commentServices';
-import { IComment } from 'models/commentSchema';
+import { IComment } from '../models/commentSchema';
+import * as UserService from '../services/userServices';
 
 
 export const createComment = async (req: Request, res: Response): Promise<void> => {
@@ -46,6 +47,25 @@ export const deleteComment = async (req: Request, res: Response): Promise<void> 
 			return;
 		}
 		console.log(' 💬 [C]*deleteComment ] commentId: ', req.params, '\nreq.user.id: ', req.user.id);
+		// verifie que Comment appartient au USer
+		const comment: IComment = await CommentService.getCommentById(commentId);
+		if (!comment) {
+			console.log(' 💬 [C]*deleteComment ] ❌ Comment not found');
+			res.status(404).json({ message: "Comment not found!" });
+			return;
+		}
+		const userId = req.user.id;
+		const user = await UserService.getUserById(userId);
+		if (!user) {
+			console.log(' 💬 [C]*deleteComment ] ❌ User not found');
+			res.status(404).json({ message: "User not found!" });
+			return;
+		}
+		if (comment.userId._id.toString() !== user._id.toString()) {
+			console.log(' 💬 [C]*deleteComment ] ❌ Unauthorized');
+			res.status(403).json({ message: "Unauthorized" });
+			return;
+		}
 		const deletedComment = await CommentService.deleteComment(commentId);
 		res.status(200).json({ message: 'Comment successfully deleted' });
