+++ backend/node_api/src/controllers/postControllers.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from 'express';
 import * as PostService from '../services/postServices';
-import { IPost } from 'models/postSchema';
+import { IPost } from '../models/postSchema';
 
 
 export const createPost = async (req: Request, res: Response): Promise<void> => {
@@ -31,6 +31,19 @@ export const deletePost = async (req: Request, res: Response): Promise<void> => 
 		const postId: string = req.params.postId;
 		if (!postId) {
 			res.status(404).json({ message: "param id: Manquant " });
+		}
+		// verifie que Post appartient au USer
+		const post: IPost = await PostService.getAPostbyPostId(postId);
+		if (!post) {
+			console.log(' 📸 🗑️ [C]*deletePost ] ❌ Post not found');
+			res.status(404).json({ message: "Post not found!" });
+			return;
+		}
+		console.log(' 📸 🗑️ [C]*deletePost ] post.userId: ', post.userId._id.toString(), ' req.user.id: ', req.user.id);
+		if (post.userId._id.toString() !== req.user.id) {
+			console.log(' 📸 🗑️ [C]*deletePost ] ❌ Unauthorized');
+			res.status(403).json({ message: "Unauthorized" });
+			return;
 		}
 		await PostService.deletePostAndRelations(postId);
 		console.log(' 📸 🗑️ [C]*deletePost ] ✅ Post deleted');
@@ -107,6 +120,32 @@ export const getAllPosts = async (req: Request, res: Response): Promise<void> =>
 		}
 	} catch (error) {
 		console.log(' 📸 [C]*getAllPosts ] ❌ Error: ', error.message);
+		res.status(500).json({ message: error.message });
+	}
+};
+
+
+export const getAllPublicPosts = async (req: Request, res: Response): Promise<void> => {
+	try {
+		console.log(' 📸 [C]*getAllPublicPosts ]... ');
+		const page: number = parseInt(req.query.page as string) || 1;
+		const limit = 5;
+		const skip = (page - 1) * limit;
+		const { posts, totalPosts, totalPages } = await PostService.getAllPublicPosts(skip, limit);
+		if (!posts) {
+			res.status(404).json({ message: " ❌ pb avec posts[] " });
+		}
+		else {
+			console.log(' 📸 [C]*getAllPublicPosts ] ✅ return Posts[]... ');
+			res.status(201).json({
+				posts,
+				totalPosts,
+				totalPages,
+				currentPage: page,
+			});
+		}
+	} catch (error) {
+		console.log(' 📸 [C]*getAllPublicPosts ] ❌ Error: ', error.message);
 		res.status(500).json({ message: error.message });
 	}
 };
